include optional platformPostUrl on RefPostData type and move RefLabel definition to types.references
