+++ app/firebase/functions/src/posts/posts.processing.ts
@@ -227,12 +227,13 @@ export class PostsProcessing {
       Array.from(Object.keys(refsLabels)).map(async (reference) => {
         const url = reference;
         const refMeta = await this.getRefMeta(url, manager, originalParsed);
-        const refPostData: RefPostData = {
+        const refPostData: RefPostData = removeUndefined({
           id: post.id,
           authorProfileId: post.authorProfileId,
           createdAtMs: post.createdAtMs,
           structuredSemantics: { labels: refsLabels[url] },
-        };
+          platformPostUrl: post.generic.thread[0].url,
+        });
         /** always delete all labels from a post for a reference */
         await this.linksService.deleteRefPost(url, postId, manager);
