+++ app/webapp/src/shared/types/types.references.ts
@@ -17,4 +17,13 @@ export interface OEmbed {
 export type RefPostData = Pick<
   AppPost,
   'id' | 'authorProfileId' | 'createdAtMs' | 'structuredSemantics'
->;
+> & {
+  platformPostUrl?: string;
+};
+
+export interface RefLabel {
+  label: string;
+  postId?: string;
+  authorProfileId?: string;
+  platformPostUrl?: string;
+}
